+++ eth/eip712/eip712_test.go
@@ -80,7 +80,7 @@ func TestEIP712TestSuite(t *testing.T) {
 func (suite *EIP712TestSuite) SetupTest() {
 	suite.config = encoding.MakeConfig(app.ModuleBasics)
 	suite.clientCtx = client.Context{}.WithTxConfig(suite.config.TxConfig)
-	suite.denom = evm.DefaultEVMDenom
+	suite.denom = evm.EVMBankDenom
 
 	sdk.GetConfig().SetBech32PrefixForAccount(appconst.AccountAddressPrefix, "")
 	eip712.SetEncodingConfig(suite.config)