+++ x/evm/genesis.go
@@ -30,7 +30,7 @@ func (gs GenesisState) Validate() error {
 	seenAccounts := make(map[string]bool)
 	for _, acc := range gs.Accounts {
 		if seenAccounts[acc.Address] {
-			return fmt.Errorf("duplicated genesis account %s", acc.Address)
+			return fmt.Errorf("duplicate genesis account %s", acc.Address)
 		}
 		if err := acc.Validate(); err != nil {
 			return fmt.Errorf("invalid genesis account %s: %w", acc.Address, err)